[FLINK-31935] Register handlers for resource requirements endpoints in all deployment modes (session, application, per-job).
