+++ flink-runtime/src/main/java/org/apache/flink/runtime/dispatcher/DispatcherGateway.java
@@ -23,7 +23,6 @@ import org.apache.flink.api.common.time.Time;
 import org.apache.flink.core.execution.SavepointFormatType;
 import org.apache.flink.runtime.clusterframework.ApplicationStatus;
 import org.apache.flink.runtime.jobgraph.JobGraph;
-import org.apache.flink.runtime.jobgraph.JobResourceRequirements;
 import org.apache.flink.runtime.messages.Acknowledge;
 import org.apache.flink.runtime.rpc.FencedRpcGateway;
 import org.apache.flink.runtime.rpc.RpcTimeout;
@@ -108,29 +107,5 @@ public interface DispatcherGateway extends FencedRpcGateway<DispatcherId>, Restf
             TriggerSavepointMode savepointMode,
             @RpcTimeout final Time timeout) {
         throw new UnsupportedOperationException();
-    }
-
-    /**
-     * Read current {@link JobResourceRequirements job resource requirements} for a given job.
-     *
-     * @param jobId job to read the resource requirements for
-     * @return Future which that contains current resource requirements.
-     */
-    default CompletableFuture<JobResourceRequirements> requestJobResourceRequirements(JobID jobId) {
-        throw new UnsupportedOperationException("Operation is not yet implemented.");
-    }
-
-    /**
-     * Update {@link JobResourceRequirements job resource requirements} for a given job. When the
-     * returned future is complete the requirements have been updated and were persisted in HA, but
-     * the job may not have been rescaled (yet).
-     *
-     * @param jobId job the given requirements belong to
-     * @param jobResourceRequirements new resource requirements for the job
-     * @return Future which is completed successfully when requirements are updated
-     */
-    default CompletableFuture<Acknowledge> updateJobResourceRequirements(
-            JobID jobId, JobResourceRequirements jobResourceRequirements) {
-        throw new UnsupportedOperationException("Operation is not yet implemented.");
     }
 }
