+++ flink-runtime/src/main/java/org/apache/flink/runtime/dispatcher/DispatcherRestEndpoint.java
@@ -26,8 +26,6 @@ import org.apache.flink.runtime.leaderelection.LeaderElectionService;
 import org.apache.flink.runtime.resourcemanager.ResourceManagerGateway;
 import org.apache.flink.runtime.rest.handler.RestHandlerConfiguration;
 import org.apache.flink.runtime.rest.handler.RestHandlerSpecification;
-import org.apache.flink.runtime.rest.handler.job.JobResourceRequirementsHandler;
-import org.apache.flink.runtime.rest.handler.job.JobResourceRequirementsUpdateHandler;
 import org.apache.flink.runtime.rest.handler.job.JobSubmitHandler;
 import org.apache.flink.runtime.rest.handler.legacy.ExecutionGraphCache;
 import org.apache.flink.runtime.rest.handler.legacy.metrics.MetricFetcher;
@@ -97,20 +95,6 @@ public class DispatcherRestEndpoint extends WebMonitorEndpoint<DispatcherGateway
                         leaderRetriever, timeout, responseHeaders, executor, clusterConfiguration);
 
         handlers.add(Tuple2.of(jobSubmitHandler.getMessageHeaders(), jobSubmitHandler));
-
-        final JobResourceRequirementsHandler jobResourceRequirementsHandler =
-                new JobResourceRequirementsHandler(leaderRetriever, timeout, responseHeaders);
-        handlers.add(
-                Tuple2.of(
-                        jobResourceRequirementsHandler.getMessageHeaders(),
-                        jobResourceRequirementsHandler));
-
-        final JobResourceRequirementsUpdateHandler jobResourceRequirementsUpdateHandler =
-                new JobResourceRequirementsUpdateHandler(leaderRetriever, timeout, responseHeaders);
-        handlers.add(
-                Tuple2.of(
-                        jobResourceRequirementsUpdateHandler.getMessageHeaders(),
-                        jobResourceRequirementsUpdateHandler));
 
         return handlers;
     }
